# packages/smartpip2/smartpip2-2.1.tar.gz/smartpip2-2.1/airflow/api_connexion/endpoints/dagsetup_endpoint.py
from airflow._vendor.connexion import NoContent
from airflow.api_connexion import security
from airflow.api_connexion.exceptions import AlreadyExists, BadRequest, NotFound
from airflow.api_connexion.schemas.dag_run_schema import (
    DAGRunCollection,
    dagrun_collection_schema,
    dagrun_schema,
    dagruns_batch_form_schema,
)
from airflow.models import DagModel, DagRun
from airflow.security import permissions
from airflow.utils.session import provide_session
from airflow.utils.types import DagRunType
from flask import current_app, g, request, jsonify
from airflow.www.app import csrf
from airflow.exceptions import AirflowException

from airflow.contrib.dagsetup.gen_airflow_dagfile import gen_airflow_dag,rm_airflow_dag,gen_airflow_config

@security.requires_access(
    [
        (permissions.ACTION_CAN_READ, permissions.RESOURCE_DAG),
        (permissions.ACTION_CAN_DELETE, permissions.RESOURCE_DAG_RUN),
    ]
)
def post_create_dagfile():
    """Create a dag file."""
    params = request.get_json(force=True)
    print(params)
    try:
        gen_airflow_dag(**params)
    except AirflowException as err:
        response = jsonify(error="{}".format(err))
        response.status_code = err.status_code
        return response
    else:
        return jsonify(msg='success create dag file')

@security.requires_access(
    [
        (permissions.ACTION_CAN_READ, permissions.RESOURCE_DAG),
        (permissions.ACTION_CAN_DELETE, permissions.RESOURCE_DAG_RUN),
    ]
)
def get_delete_dagfile():
    """delete the dag."""
    try:
        name = request.args.get('name')
        rm_airflow_dag(name)
    except AirflowException as err:
        response = jsonify(error="{}".format(err))
        response.status_code = err.status_code
        return response
    else:
        return jsonify(msg= f'success delete dag file {name}')

@security.requires_access(
    [
        (permissions.ACTION_CAN_READ, permissions.RESOURCE_DAG),
        (permissions.ACTION_CAN_DELETE, permissions.RESOURCE_DAG_RUN),
    ]
)
def get_gen_dagconfig():
    """generate the dag config base on etl script."""
    try:
        name = request.args.get('name')
        dag_str = gen_airflow_config(name)
    except AirflowException as err:
        response = jsonify(error="{}".format(err))
        response.status_code = err.status_code
        return response
    else:
        return jsonify(msg= dag_str)